+++ src/utils/transaction.ts
@@ -2,7 +2,7 @@ import {
   Address,
   Hex,
   bytesToHex,
-  hexToNumber,
+  hexToBytes,
   keccak256,
   parseTransaction,
   recoverPublicKey,
@@ -12,36 +12,13 @@ import {
 import { TransactionWithSignature } from "../types";
 import { FeeMarketEIP1559Transaction } from "@ethereumjs/tx";
 
-export function hexStringToUint8Array(hexString: string): Uint8Array {
-  // Remove the "0x" prefix if it's present
-  hexString = hexString.replace(/^0x/, "");
-
-  // Ensure the hex string has an even length
-  if (hexString.length % 2 !== 0) {
-    console.error(
-      "Hex string has an odd length, which is not valid for byte representation."
-    );
-    return new Uint8Array();
-  }
-
-  // Create a Uint8Array with the length of half the hex string length
-  const byteArray = new Uint8Array(hexString.length / 2);
-
-  // Convert each hex byte (two hex characters) to a decimal value and fill the Uint8Array
-  for (let i = 0, j = 0; i < hexString.length; i += 2, j++) {
-    byteArray[j] = parseInt(hexString.substring(i, i + 2), 16);
-  }
-
-  return byteArray;
-}
-
 export function ethersJsAddSignature(
   tx: TransactionWithSignature,
   sender: Address
 ): Hex {
   const { transaction: unsignedTxHash, signature: sig } = tx;
   const transaction = FeeMarketEIP1559Transaction.fromSerializedTx(
-    hexStringToUint8Array(unsignedTxHash)
+    hexToBytes(unsignedTxHash)
   );
   const r = Buffer.from(sig.big_r.substring(2), "hex");
   const s = Buffer.from(sig.big_s, "hex");
@@ -83,8 +60,6 @@ export async function viemAddSig(
       hash: serializeTransaction(tx),
       signature,
     });
-    const v = hexToNumber(`0x${signature.slice(130)}`);
-    console.log("V Value:", v);
     return pk.toString().toLowerCase() === sender.toLowerCase();
   });
   if (!signature) {
@@ -106,5 +81,5 @@ export async function buildTxPayload(
 ): Promise<number[]> {
   // Compute the Transaction Message Hash.
   const messageHash = await keccak256(unsignedTxHash);
-  return Array.from(hexStringToUint8Array(messageHash).slice().reverse());
+  return Array.from(hexToBytes(messageHash).slice().reverse());
 }
