+++ tests/e2e.test.ts
@@ -1,5 +1,12 @@
-describe("End To End", () => {
-  it.skip("Runs the Send ETH Tx", () => {
-    expect(1).toEqual(0);
+import { setupNearEthAdapter } from "../examples/setup";
+
+describe.skip("End To End", () => {
+  it("Runs the Send ETH Tx", async () => {
+    const evm = await setupNearEthAdapter();
+    await evm.signAndSendTransaction({
+      to: "0xdeADBeeF0000000000000000000000000b00B1e5",
+      // THIS IS ONE WEI!
+      value: 1n,
+    });
   });
 });
